src/auth-svc/requirements.txt
flask 
gunicorn
python-dotenv
Flask-SQLAlchemy
Flask-Migrate
PyJWT
psycopg2-binary
bcrypt